backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Pillow==10.1.0
dj-database-url==2.1.0
gunicorn==21.2.0
whitenoise==6.6.0
django-storages==1.14.2
boto3==1.34.11 